% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workshop_certificate.R
\name{create_workshop_certificates}
\alias{create_workshop_certificates}
\title{Create certificates for all attendees}
\usage{
create_workshop_certificates(date, workshop, curriculum, certifyer, credentials,
  attendees, dir)
}
\arguments{
\item{date}{Date of the workshop, as a date.}

\item{workshop}{Workshop title, character.}

\item{curriculum}{Path to the workshop curriculum (.md), character.}

\item{certifyer}{Person certifying, character.}

\item{credentials}{Credentials of the certifying person, character.}

\item{attendees}{Names of attendees, character vector.}

\item{dir}{Directory where to output the pdf certificates, character.}
}
\description{
Create certificates for all attendees
}
\examples{
\dontrun{
# Fake names generated via charlatan::ch_name
attendees <- c("Marnie Dickinson", "Dr. Marlin Wilderman")
date <- structure(17532, class = "Date")
workshop <- "Package development workshop"
curriculum <- system.file("rmarkdown", "templates",
"workshop_certificate", "resources",
"default_workshop_contents.md", package = "fwdbrand")
certifyer <- "Zaire Crooks"
credentials <- "Forwards teaching team member"
dir <- "certificates"
create_workshop_certificates(date, workshop, curriculum,
 certifyer,
credentials,
attendees,
dir)
}
}
